requirements.txt
anyascii==0.3.1
appdirs==1.4.4
asttokens==2.2.1
audioread==3.0.0
Babel==2.11.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coqpit==0.0.17
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.7
decorator==5.1.1
docopt==0.6.2
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
fsspec==2023.1.0
future==0.18.3
g2pkk==0.1.2
greenlet==2.0.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
inflect==5.6.0
ipython==8.9.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonlines==1.2.0
kiwisolver==1.4.4
librosa==0.8.0
llvmlite==0.38.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mecab-python3==1.0.5
more-itertools==9.0.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.55.1
numpy==1.21.6
openai-whisper==20230124
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
playwright==1.30.0
pooch==1.6.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pyee==9.0.4
Pygments==2.14.0
pynndescent==0.5.8
pyparsing==3.0.9
pypinyin==0.48.0
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
resampy==0.4.2
scikit-learn==1.2.1
scipy==1.10.0
simpleaudio==1.0.4
six==1.16.0
soundfile==0.11.0
SpeechRecognition==3.9.0
stack-data==0.6.2
tensorboardX==2.5.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
tqdm==4.64.1
trainer==0.0.20
traitlets==5.9.0
transformers==4.26.0
TTS==0.10.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
zipp==3.12.0
